Rename `history_processor.py` module to `process_history.py`
Match the file-name convention used by `process_event_stream.py`, per
review feedback on #5197.
