+++ pydantic_ai_slim/pydantic_ai/capabilities/__init__.py
@@ -18,10 +18,6 @@ from .abstract import (
 from .builtin_or_local import BuiltinOrLocalTool
 from .builtin_tool import BuiltinTool
 from .combined import CombinedCapability
-from .history_processor import (
-    HistoryProcessor,  # pyright: ignore[reportDeprecated]
-    ProcessHistory,
-)
 from .hooks import Hooks, HookTimeoutError
 from .image_generation import ImageGeneration
 from .include_return_schemas import IncludeToolReturnSchemas
@@ -29,6 +25,10 @@ from .mcp import MCP
 from .prefix_tools import PrefixTools
 from .prepare_tools import PrepareTools
 from .process_event_stream import ProcessEventStream
+from .process_history import (
+    HistoryProcessor,  # pyright: ignore[reportDeprecated]
+    ProcessHistory,
+)
 from .reinject_system_prompt import ReinjectSystemPrompt
 from .set_tool_metadata import SetToolMetadata
 from .thinking import Thinking
